refactoring; add FixedPoint
